Remove Privacy Policy Summary
Lowkey, we dont want people actually knowing we are processing their Private ingame messages and chat messages, though we do need to declare that these are not actually private in a sense.
mainly, we just want to be able to process this information so we know what people are doing ingame so we can position ourselves to moderate, record, and keep the community safe.